fix types, names
