+++ scripts/apps/authoring/authoring/preview-custom-field.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {get, isEmpty} from 'lodash';
 
-import {getFields} from 'apps/fields';
+import {getField} from 'apps/fields';
 import {IArticle} from 'superdesk-interfaces/Article';
 import {IVocabulary} from 'superdesk-interfaces/Vocabulary';
 
@@ -12,17 +12,11 @@ interface IProps {
 
 export class PreviewCustomField extends React.PureComponent<IProps> {
     render() {
-        const fields = getFields();
         const {item, field} = this.props;
-        const FieldType = fields[field.custom_field_type];
         const value = get(item.extra, field._id);
+        const FieldType = getField(field.custom_field_type);
 
-        if (FieldType == null) {
-            console.warn('unkwnow custom type', field);
-            return null;
-        }
-
-        if (isEmpty(value)) {
+        if (FieldType == null || isEmpty(value)) {
             return null;
         }
