+++ scripts/apps/workspace/content/services/ContentService.ts
@@ -324,7 +324,7 @@ export function ContentService(api, superdesk, templates, desks, packages, archi
             self._fieldsPromise = api.getAll('vocabularies', {
                 where: {
                     $or: [
-                        {field_type: {$in: ['text', 'date', 'media', 'embed', 'urls']}},
+                        {field_type: {$in: ['text', 'date', 'media', 'embed', 'urls', 'custom']}},
                         {service: {$exists: true}},
                     ],
                 },
