+++ scripts/core/ui/components/alert.tsx
@@ -1,18 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {IPropsComponentAlert} from 'superdesk-api';
+import {IAlertComponentProps} from 'superdesk-api';
+import {assertNever} from 'core/helpers/typescript-helpers';
 
-const CLASSNAMES = {
-    info: 'sd-alert--primary',
-    warning: 'sd-alert--warning',
-    error: 'sd-alert--alert',
-};
+function getClassName(alertType: IAlertComponentProps['type']) {
+    switch (alertType) {
+    case 'info':
+        return 'sd-alert--primary';
+    case 'error':
+        return 'sd-alert--alert';
+    case 'warning':
+        return 'sd-alert--warning';
+    default:
+        assertNever(alertType);
+    }
+}
 
-export const Alert: React.StatelessComponent<IPropsComponentAlert> = (props) => {
+export const Alert: React.StatelessComponent<IAlertComponentProps> = (props) => {
     const className = [
         'sd-alert',
         props.hollow ? 'sd-alert--hollow' : '',
-        CLASSNAMES[props.type],
+        getClassName(props.type),
     ].join(' ');
 
     return (
@@ -23,5 +31,5 @@ export const Alert: React.StatelessComponent<IPropsComponentAlert> = (props) => 
 Alert.propTypes = {
     type: PropTypes.oneOf(['info', 'warning', 'error']),
     hollow: PropTypes.bool,
-    children: PropTypes.node.isRequired,
+    children: PropTypes.node,
 };
